pyproject.toml
[project]
name = "md2docx-mcp-server"
version = "1.3.0"
description = "A unified MCP server for Markdown to DOCX, PPTX, and LaTeX/PDF conversion with self-maintained converters"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.10.1",
    "markdown-it-py",
    "python-docx",
    "python-pptx>=0.6.21",
    "lxml>=4.9.0",
    "pillow>=10.0.0",
    "requests",
    "html2docx",
    "aiofiles>=23.0.0",
    # MD2LaTeX dependencies
    "mistune>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black",
    "pyinstaller",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "core*"]
